Change ccontroller

+++ app/controllers/articles_controller.rb
@@ -20,12 +20,12 @@ class ArticlesController < ApplicationController
 
   def destroy
     article = Article.find(params[:id])
-    article.destroy
+    article.destroy if article.user_id == current_user.id
     redirect_to action: :index
   end
 
   def update
-    @article.update(article_params)
+    @article.update(article_params) if@article.user_id == current_user.id
     redirect_to action: :index
   end
